dom0 plays should be executed on current host
Setting dom0 as the target host for qubesos module actions makes sense
and should behave the same as localhost.
This commit skips proxy execution when the target host is dom0 (as it already does for localhost).
Inventory creation is also modified to set local connection as default for dom0 host.